chore: move tests and test resources to `kafka/listeners/asn1` for correct organization
