+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/kafka/listeners/asn1/RawEncodedSPATJsonRouterTest.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.kafka.listeners;
+package us.dot.its.jpo.ode.kafka.listeners.asn1;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -22,7 +22,6 @@ import org.springframework.test.context.ContextConfiguration;
 import us.dot.its.jpo.ode.config.SerializationConfig;
 import us.dot.its.jpo.ode.kafka.KafkaConsumerConfig;
 import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
-import us.dot.its.jpo.ode.kafka.listeners.asn1.RawEncodedSRMJsonRouter;
 import us.dot.its.jpo.ode.kafka.producer.KafkaProducerConfig;
 import us.dot.its.jpo.ode.kafka.topics.Asn1CoderTopics;
 import us.dot.its.jpo.ode.kafka.topics.RawEncodedJsonTopics;
@@ -32,13 +31,13 @@ import us.dot.its.jpo.ode.udp.controller.UDPReceiverProperties;
 @SpringBootTest(
     classes = {
         KafkaProducerConfig.class,
+        RawEncodedSPATJsonRouter.class,
         KafkaConsumerConfig.class,
-        RawEncodedSRMJsonRouter.class,
         SerializationConfig.class
     },
     properties = {
-        "ode.kafka.topics.raw-encoded-json.srm=topic.Asn1DecoderTestSRMJSON",
-        "ode.kafka.topics.asn1.decoder-input=topic.Asn1DecoderSRMInput"
+        "ode.kafka.topics.raw-encoded-json.spat=topic.Asn1DecoderTestSPATJSON",
+        "ode.kafka.topics.asn1.decoder-input=topic.Asn1DecoderSPATInput"
     })
 @EnableConfigurationProperties
 @ContextConfiguration(classes = {
@@ -46,22 +45,23 @@ import us.dot.its.jpo.ode.udp.controller.UDPReceiverProperties;
     RawEncodedJsonTopics.class, KafkaProperties.class, Asn1CoderTopics.class
 })
 @DirtiesContext
-class RawEncodedSRMJsonRouterTest {
+class RawEncodedSPATJsonRouterTest {
 
   @Autowired
   Asn1CoderTopics asn1CoderTopics;
   @Autowired
   RawEncodedJsonTopics rawEncodedJsonTopics;
   @Autowired
-  private KafkaTemplate<String, String> kafkaTemplate;
+  KafkaTemplate<String, String> kafkaTemplate;
 
   @Test
   void testListen() throws JSONException, IOException {
     var embeddedKafka = EmbeddedKafkaHolder.getEmbeddedKafka();
-    EmbeddedKafkaHolder.addTopics(asn1CoderTopics.getDecoderInput(), rawEncodedJsonTopics.getSrm());
+    EmbeddedKafkaHolder.addTopics(asn1CoderTopics.getDecoderInput(),
+        rawEncodedJsonTopics.getSpat());
 
     Map<String, Object> consumerProps =
-        KafkaTestUtils.consumerProps("Asn1DecodeSRMJSONTestConsumer", "false", embeddedKafka);
+        KafkaTestUtils.consumerProps("Asn1DecodeSPATJSONTestConsumer", "false", embeddedKafka);
     var cf =
         new DefaultKafkaConsumerFactory<>(consumerProps,
             new StringDeserializer(), new StringDeserializer());
@@ -70,17 +70,19 @@ class RawEncodedSRMJsonRouterTest {
 
     var classLoader = getClass().getClassLoader();
     InputStream inputStream = classLoader
-        .getResourceAsStream("us/dot/its/jpo/ode/services/asn1/messages/decoder-input-srm.json");
+        .getResourceAsStream(
+            "us/dot/its/jpo/ode/kafka/listeners/asn1/messages/decoder-input-spat.json");
     assert inputStream != null;
-    var json = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
-    kafkaTemplate.send(rawEncodedJsonTopics.getSrm(), json);
+    var spatJson = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+    kafkaTemplate.send(rawEncodedJsonTopics.getSpat(), spatJson);
+
+    var consumedSpat =
+        KafkaTestUtils.getSingleRecord(testConsumer, asn1CoderTopics.getDecoderInput());
 
     inputStream = classLoader
-        .getResourceAsStream("us/dot/its/jpo/ode/services/asn1/messages/expected-srm.xml");
+        .getResourceAsStream("us/dot/its/jpo/ode/kafka/listeners/asn1/messages/expected-spat.xml");
     assert inputStream != null;
-    var expectedSrm = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
-
-    var producedSrm = KafkaTestUtils.getSingleRecord(testConsumer, asn1CoderTopics.getDecoderInput());
-    assertEquals(expectedSrm, producedSrm.value());
+    var expectedSpat = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+    assertEquals(expectedSpat, consumedSpat.value());
   }
 }
