+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/kafka/listeners/asn1/RawEncodedSSMJsonRouterTest.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.kafka.listeners;
+package us.dot.its.jpo.ode.kafka.listeners.asn1;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -11,7 +11,6 @@ import org.apache.kafka.common.serialization.StringDeserializer;
 import org.json.JSONException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -23,8 +22,8 @@ import org.springframework.test.context.ContextConfiguration;
 import us.dot.its.jpo.ode.config.SerializationConfig;
 import us.dot.its.jpo.ode.kafka.KafkaConsumerConfig;
 import us.dot.its.jpo.ode.kafka.OdeKafkaProperties;
-import us.dot.its.jpo.ode.kafka.listeners.asn1.RawEncodedBSMJsonRouter;
 import us.dot.its.jpo.ode.kafka.producer.KafkaProducerConfig;
+import us.dot.its.jpo.ode.kafka.topics.Asn1CoderTopics;
 import us.dot.its.jpo.ode.kafka.topics.RawEncodedJsonTopics;
 import us.dot.its.jpo.ode.test.utilities.EmbeddedKafkaHolder;
 import us.dot.its.jpo.ode.udp.controller.UDPReceiverProperties;
@@ -33,58 +32,55 @@ import us.dot.its.jpo.ode.udp.controller.UDPReceiverProperties;
     classes = {
         KafkaProducerConfig.class,
         KafkaConsumerConfig.class,
-        RawEncodedBSMJsonRouter.class,
+        RawEncodedSSMJsonRouter.class,
         SerializationConfig.class
     },
     properties = {
-        "ode.kafka.topics.raw-encoded-json.bsm=topic.Asn1DecoderTestBSMJSON",
-        "ode.kafka.topics.asn1.decoder-input=topic.Asn1DecoderBSMInput"
+        "ode.kafka.topics.raw-encoded-json.ssm=topic.Asn1DecoderTestSSMJSON",
+        "ode.kafka.topics.asn1.decoder-input=topic.Asn1DecoderSSMInput"
     })
 @EnableConfigurationProperties
 @ContextConfiguration(classes = {
     UDPReceiverProperties.class, OdeKafkaProperties.class,
-    RawEncodedJsonTopics.class, KafkaProperties.class
+    RawEncodedJsonTopics.class, KafkaProperties.class, Asn1CoderTopics.class
 })
 @DirtiesContext
-class RawEncodedBSMJsonRouterTest {
-
-  @Value(value = "${ode.kafka.topics.raw-encoded-json.bsm}")
-  private String rawEncodedBsmJson;
-
-  @Value(value = "${ode.kafka.topics.asn1.decoder-input}")
-  private String asn1DecoderInput;
+class RawEncodedSSMJsonRouterTest {
 
   @Autowired
-  KafkaTemplate<String, String> kafkaTemplate;
+  Asn1CoderTopics asn1CoderTopics;
+  @Autowired
+  RawEncodedJsonTopics rawEncodedJsonTopics;
+  @Autowired
+  private KafkaTemplate<String, String> kafkaTemplate;
 
   @Test
   void testListen() throws JSONException, IOException {
     var embeddedKafka = EmbeddedKafkaHolder.getEmbeddedKafka();
-    EmbeddedKafkaHolder.addTopics(asn1DecoderInput, rawEncodedBsmJson);
+    EmbeddedKafkaHolder.addTopics(asn1CoderTopics.getDecoderInput(), rawEncodedJsonTopics.getSsm());
 
     Map<String, Object> consumerProps =
-        KafkaTestUtils.consumerProps("Asn1DecodeBSMJSONTestConsumer", "false", embeddedKafka);
+        KafkaTestUtils.consumerProps("Asn1DecodeSSMJSONTestConsumer", "false", embeddedKafka);
     var cf =
         new DefaultKafkaConsumerFactory<>(consumerProps,
             new StringDeserializer(), new StringDeserializer());
     Consumer<String, String> testConsumer = cf.createConsumer();
-    embeddedKafka.consumeFromAnEmbeddedTopic(testConsumer, asn1DecoderInput);
+    embeddedKafka.consumeFromAnEmbeddedTopic(testConsumer, asn1CoderTopics.getDecoderInput());
 
     var classLoader = getClass().getClassLoader();
     InputStream inputStream = classLoader
-        .getResourceAsStream("us/dot/its/jpo/ode/services/asn1/messages/decoder-input-bsm.json");
+        .getResourceAsStream(
+            "us/dot/its/jpo/ode/kafka/listeners/asn1/messages/decoder-input-ssm.json");
     assert inputStream != null;
-    var bsmJson = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
-    kafkaTemplate.send(rawEncodedBsmJson, bsmJson);
+    var json = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+    kafkaTemplate.send(rawEncodedJsonTopics.getSsm(), json);
 
     inputStream = classLoader
-        .getResourceAsStream("us/dot/its/jpo/ode/services/asn1/messages/expected-bsm.xml");
+        .getResourceAsStream("us/dot/its/jpo/ode/kafka/listeners/asn1/messages/expected-ssm.xml");
     assert inputStream != null;
-    var expectedBsm = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
+    var expectedSSM = new String(inputStream.readAllBytes(), StandardCharsets.UTF_8);
 
-    var produced =
-        KafkaTestUtils.getSingleRecord(testConsumer, asn1DecoderInput);
-    var odeBsmData = produced.value();
-    assertEquals(expectedBsm, odeBsmData);
+    var consumedSSM = KafkaTestUtils.getSingleRecord(testConsumer, asn1CoderTopics.getDecoderInput());
+    assertEquals(expectedSSM, consumedSSM.value());
   }
 }
