+++ jpo-ode-svcs/src/test/resources/us/dot/its/jpo/ode/kafka/listeners/asn1/messages/expected-ssm.xml
